Python/requirements.txt
Flask
pymongo
pandas
matplotlib
scikit-learn
python-dotenv
